chatbot/requirements.txt
# Core
python-dotenv==1.1.0
google-adk>=1.19.0
uvicorn==0.34.3
fastapi==0.115.12

# Database
psycopg2-binary==2.9.10
asyncpg==0.30.0
greenlet>=3.0.0
sqlalchemy>=2.0.0

# HTTP client
aiohttp>=3.9.0

# Auth
PyJWT==2.10.1
